feat: develop category and subCategory
